# Particle property pool improvements