--- a/include/deal.II/particles/property_pool.h
+++ b/include/deal.II/particles/property_pool.h
@@ -20,8 +20,6 @@
 
 #include <deal.II/base/array_view.h>
 
-#include <set>
-
 
 DEAL_II_NAMESPACE_OPEN
 
@@ -54,7 +52,7 @@ namespace Particles
      * uniquely identifies the slot of memory that is reserved for this
      * particle.
      */
-    using Handle = double *;
+    using Handle = unsigned int;
 
     /**
      * Define a default (invalid) value for handles.
@@ -74,20 +72,33 @@ namespace Particles
     ~PropertyPool();
 
     /**
-     * Return a new handle that allows accessing the reserved block
-     * of memory. If the number of properties is zero this will return an
-     * invalid handle.
+     * Clear the dynamic memory allocated by this class. This function
+     * ensures that all memory that had previously been allocated using
+     * allocate_properties_array() has also been returned via
+     * deallocate_properties_array().
+     */
+    void
+    clear();
+
+    /**
+     * Return a new handle that allows accessing the reserved particle
+     * properties. If the number of properties is zero this will return an
+     * invalid handle. Handles can be copied, but after deallocating one
+     * of the copied handles using any of the other copies will cause
+     * undefined behavior.
      */
     Handle
     allocate_properties_array();
 
     /**
      * Mark the properties corresponding to the handle @p handle as
-     * deleted. Calling this function more than once for the same
-     * handle causes undefined behavior.
+     * deleted and invalidate the @p handle. Calling this function
+     * more than once for the same handle (e.g. by calling it again
+     * with a copy of the previously invalidated handle) causes
+     * undefined behavior.
      */
     void
-    deallocate_properties_array(const Handle handle);
+    deallocate_properties_array(Handle &handle);
 
     /**
      * Return an ArrayView to the properties that correspond to the given
@@ -116,12 +127,42 @@ namespace Particles
     const unsigned int n_properties;
 
     /**
-     * A collection of handles that have been created by
-     * allocate_properties_array() and have not been destroyed by
-     * deallocate_properties_array().
+     * The currently allocated properties (whether assigned to
+     * a particle or available for assignment).
      */
-    std::set<Handle> currently_open_handles;
+    std::vector<double> properties;
+
+    /**
+     * A collection of handles that have been created by
+     * allocate_properties_array() and have been destroyed by
+     * deallocate_properties_array(). Since the memory is still
+     * allocated these handles can be reused for new particles
+     * to avoid memory allocation.
+     */
+    std::vector<Handle> currently_available_handles;
   };
+
+  /* ---------------------- inline and template functions ------------------ */
+
+  inline ArrayView<double>
+  PropertyPool::get_properties(const Handle handle)
+  {
+    const std::vector<double>::size_type data_index =
+      (handle != invalid_handle) ? handle * n_properties : 0;
+
+    // Ideally we would need to assert that 'handle' has not been deallocated
+    // by searching through 'currently_available_handles'. However, this
+    // is expensive and this function is performance critical, so instead
+    // just check against the properties range, and rely on the fact
+    // that handles are invalidated when handed over to
+    // deallocate_properties_array().
+    Assert(data_index <= properties.size() - n_properties,
+           ExcMessage("Invalid property handle. This can happen if the "
+                      "handle was duplicated and then one copy was deallocated "
+                      "before trying to access the properties."));
+
+    return ArrayView<double>(properties.data() + data_index, n_properties);
+  }
 
 
 } // namespace Particles