This is the very first step towards #11206 and also limits the performance impact of #11059. It combines a few changes that only make sense together. It does:
1. Allocate particle properties in a common vector, instead of one array per particle. This speeds up allocation and removes the need for deallocation for every particle.
2. Instead of tracking every open handle, it only tracks handles that have been allocated and returned (i.e. they are available for reuse). It is assumed that all handles that have not been returned are still in use. To check whether all handles have been returned we can now check the number of available handles against the size of the properties vector.
3. Instead of a std::set of handles it tracks a std::vector. This offers a huge performance boost (larger than the actual allocation/deallocation itself), with the minor drawback that now it is expensive to check if a particular handle has been returned (we would need to search through the vector of available handles).

I started out only wanting to implement 1., but in my benchmark I noticed that the tracking of handles is very expensive. I used a benchmark program for the timing below that first allocates 1e7 particle properties, then iterates over them, then removes some (two times 25% of the total number to mess up the order in the available handles), then reallocate properties up to the full size of 1e7, then iterate again (to measure iteration over unsorted properties), then deallocate everything. Everything is happening in release mode. This PR is more than 10X faster than master, and it is even faster than master without tracking the handles (i.e. before #11059). It also speeds up iterating over the particle properties (as long as each particle only has a few properties and they are roughly sorted) because memory access is not longer as fragmented.

I still need to do some testing, and there are some limitations, but I do not know when I can get back to this, so I wanted to ask for opinions at this point.

Limitations:
1. ~~No way to reduce the memory allocation as of yet. The PropertyPool will require as much memory as it needed for the maximum number of particles during its lifetime, so it is not growing indefinitely, but it may get large.~~ Edit: Memory will be reduced and reallocated whenever ParticleHandler::clear_particles is called.
2. ~~I need to test serialization more.~~ Tests seem to work.
3. No way of sorting properties in order ~~after they got allocatet as of yet~~ except for copying the whole particle handler or during mesh refinement.

Master:
```
+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      41.5s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| Property allocation             |         1 |      19.6s |        47% |
| Property deallocation 1, 25%    |         1 |      3.95s |       9.5% |
| Property deallocation 2, 18.75% |         1 |      3.09s |       7.5% |
| Property deallocation, all      |         1 |      7.51s |        18% |
| Property iteration, sorted      |         1 |     0.269s |      0.65% |
| Property iteration, unsorted    |         1 |     0.391s |      0.94% |
| Property reallocation, 43.75%   |         1 |       5.9s |        14% |
+---------------------------------+-----------+------------+------------+
```
Master without tracking the handles:
```
+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |       5.7s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| Property allocation             |         1 |       1.8s |        32% |
| Property deallocation 1, 25%    |         1 |     0.889s |        16% |
| Property deallocation 2, 18.75% |         1 |     0.882s |        15% |
| Property deallocation, all      |         1 |     0.714s |        13% |
| Property iteration, sorted      |         1 |     0.142s |       2.5% |
| Property iteration, unsorted    |         1 |     0.259s |       4.5% |
| Property reallocation, 43.75%   |         1 |     0.879s |        15% |
+---------------------------------+-----------+------------+------------+
```
This PR (tracking available handles, vector storage for handles, vector storage for properties):
```
+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      2.99s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| Property allocation             |         1 |     0.606s |        20% |
| Property deallocation 1, 25%    |         1 |      0.79s |        26% |
| Property deallocation 2, 18.75% |         1 |     0.795s |        27% |
| Property deallocation, all      |         1 |     0.217s |       7.3% |
| Property iteration, sorted      |         1 |    0.0901s |         3% |
| Property iteration, unsorted    |         1 |     0.115s |       3.9% |
| Property reallocation, 43.75%   |         1 |     0.219s |       7.3% |
+---------------------------------+-----------+------------+------------+
```