@gassmoeller 2020-12-04
Yes that is correct. The one situation that could be problematic is if one MPI rank holds a lot of particles at a particular point (e.g. all particles are generated on one process and then distributed, or particles are clustered in a part of the domain and move together). I updated the PR with two additional safeguards against that. 1. Now if the last handle is deallocated the properties and the handles vector is resized to 0. This will not deallocate the memory, but it will mean if new particles are generated their properties will be stored continuously. 2. I added a PropertyPool::reinit() function that releases all memory (only callable if no handle is owned). This function is called whenever ParticleHandler::clear_particles is called, which happens for example during mesh refinement.

I think this is ready from my side, but would need a few more eyes on the changes.