@gassmoeller 2020-12-10
Great points, I think I addressed your comments. I chose `types::global_dof_index` for the Handle type, because that one depends on the 64bit cmake setting. Vertex index seems to be always 64 bit, which would seem a bit excessive (although I have not measured if it would actually make a difference in performance). But when moving the  `get_properties` function I wondered if the correct type for the handle should not be `std::vector<double>::size_type`? Or is that tying the type of the handle to close to our internal storage structure? Maybe I am overthinking this. 
Otherwise ready on my side.