@bangerth 2020-12-11
The "correct" data type for the handle is `std::vector<>::size_type` because, strictly speaking, that's exactly what is used for indexing. But, that also requires the `Particle` class to store an 8-byte data type when it would be nice to use only 4 bytes. At the same time, let's pretend you're only storing a 4-byte `unsigned int`, than the maximal size of an array you can store is already 32 GB given that each entry of that array is at least a `double`. Moreover, if you made that not an index into the array, but an index into "chunks of size n_properties", you'd get a multiple.

I think the likelihood that we want more than 4B particles on one MPI rank is pretty small. It wouldn't bother me if we just took an `unsigned int` and had an assert somewhere that aborts the program if someone tried to create more than 4B particles/required more than 4B unique handles.