@kronbichler 2020-12-11
> At the same time, let's pretend you're only storing a 4-byte `unsigned int`

I would be fine with using unsigned ints as the handle type. Given that we use a typedef we could later change it (modulo some possible bugs related to the hard-coded variant) without interfering with user codes.