@gassmoeller 2020-12-12
I changed the initialization of the invalid handle.

> Actually, what you're still missing here is an Assert(handle != invalid_handle, ...);. Nothing good can come off presenting an invalid handle here. If you add that assertion, you don't need the ?: expression.

The problem here is that we also return an invalid handle in `allocate_properties` if a particle has 0 properties. So `get_properties(invalid_handle)` could either mean the properties have already been allocated (in which case we should throw) or simply that `n_properties` is zero (in which case we should return an empty ArrayView). It is maybe not ideal design, but I am worried that changing that behavior will require further changes in the Particle class (e.g. a particle could use `properties == invalid_handle` to determine if it has properties, instead of `get_properties().size`). We should probably disentangle this, but maybe in a follow-up PR? We will have to touch this again anyway, when we want particles to store all their members inside the property pool, because then all particles will have to have properties all the time.