@bangerth 2020-12-12
Let's get this merged. I have a couple of follow-up patches lined up that also start moving the locations and reference locations of particles into the property pool.